initial FilterButton.jsx component
